feat: insert memo list to script response

+++ src/script/script.service.ts
@@ -196,9 +196,10 @@ export class ScriptService {
 
     // 메모 생성
     async createMemo(createMemoDto: CreateMemoDto): Promise<ReturnScriptDto> {
-      const script: Script = await this.checkScriptOwner(createMemoDto.userId, createMemoDto.script.id);
-      const memo: Memo = await this.memoRepository.createMemo(createMemoDto);
-      // const script: Script = await this.scriptRepository()
+      const scriptId: number = createMemoDto.script.id;
+      await this.checkScriptOwner(createMemoDto.userId, createMemoDto.script.id);
+      await this.memoRepository.createMemo(createMemoDto);
+      const script: Script = await this.scriptRepository.findOneOrFail(scriptId)
       const returnScriptDto: ReturnScriptDto = new ReturnScriptDto(script);
       return returnScriptDto;
     }
@@ -209,8 +210,9 @@ export class ScriptService {
       const scriptId: number = deleteMemoDto.scriptId;
       const memoId: number = deleteMemoDto.memoId;
 
-      const script: Script = await this.checkScriptOwner(userId, scriptId);
+      await this.checkScriptOwner(userId, scriptId);
       await this.memoRepository.deleteMemo(memoId);
+      const script: Script = await this.scriptRepository.findOneOrFail(scriptId)
       const returnScriptDto: ReturnScriptDto = new ReturnScriptDto(script);
       return returnScriptDto;
     }
